+++ finality-provider/service/fp_instance_test.go
@@ -2,24 +2,25 @@ package service_test
 
 import (
 	"fmt"
-	bbntypes "github.com/babylonlabs-io/babylon/types"
-	"github.com/golang/mock/gomock"
-	"github.com/stretchr/testify/require"
-	"go.uber.org/zap"
 	"math/rand"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/babylonlabs-io/babylon/testutil/datagen"
+	bbntypes "github.com/babylonlabs-io/babylon/types"
 	ftypes "github.com/babylonlabs-io/babylon/x/finality/types"
+	"github.com/golang/mock/gomock"
+	"github.com/stretchr/testify/require"
+	"go.uber.org/zap"
 
 	"github.com/babylonlabs-io/finality-provider/clientcontroller"
 	"github.com/babylonlabs-io/finality-provider/eotsmanager"
 	eotscfg "github.com/babylonlabs-io/finality-provider/eotsmanager/config"
 	"github.com/babylonlabs-io/finality-provider/finality-provider/config"
-	"github.com/babylonlabs-io/finality-provider/finality-provider/proto"
 	"github.com/babylonlabs-io/finality-provider/finality-provider/service"
+	fpkr "github.com/babylonlabs-io/finality-provider/keyring"
 	"github.com/babylonlabs-io/finality-provider/metrics"
 	"github.com/babylonlabs-io/finality-provider/testutil"
 	"github.com/babylonlabs-io/finality-provider/types"
@@ -176,14 +177,33 @@ func startFinalityProviderAppWithRegisteredFp(
 	require.NoError(t, err)
 	eotsPk, err := bbntypes.NewBIP340PubKey(eotsPkBz)
 	require.NoError(t, err)
-	fp := testutil.GenStoredFinalityProvider(r, t, app, passphrase, hdPath, eotsPk)
 	pubRandProofStore := app.GetPubRandProofStore()
 	fpStore := app.GetFinalityProviderStore()
-	err = fpStore.SetFpStatus(fp.BtcPk, proto.FinalityProviderStatus_REGISTERED)
+	keyName := datagen.GenRandomHexStr(r, 10)
+	chainID := datagen.GenRandomHexStr(r, 10)
+	input := strings.NewReader("")
+	kr, err := fpkr.CreateKeyring(
+		fpCfg.BabylonConfig.KeyDirectory,
+		fpCfg.BabylonConfig.ChainID,
+		fpCfg.BabylonConfig.KeyringBackend,
+		input,
+	)
 	require.NoError(t, err)
-	// TODO: use mock metrics
+	kc, err := fpkr.NewChainKeyringControllerWithKeyring(kr, keyName, input)
+	require.NoError(t, err)
+	keyInfo, err := kc.CreateChainKey("", "", "")
+	require.NoError(t, err)
+	fpAddr := keyInfo.AccAddress
+	err = fpStore.CreateFinalityProvider(
+		fpAddr,
+		eotsPk.MustToBTCPK(),
+		testutil.RandomDescription(r),
+		testutil.ZeroCommissionRate(),
+		chainID,
+	)
+	require.NoError(t, err)
 	m := metrics.NewFpMetrics()
-	fpIns, err := service.NewFinalityProviderInstance(fp.GetBIP340BTCPK(), &fpCfg, fpStore, pubRandProofStore, cc, em, m, passphrase, make(chan *service.CriticalError), logger)
+	fpIns, err := service.NewFinalityProviderInstance(eotsPk, &fpCfg, fpStore, pubRandProofStore, cc, em, m, passphrase, make(chan *service.CriticalError), logger)
 	require.NoError(t, err)
 
 	cleanUp := func() {