+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#179](https://github.com/babylonlabs-io/finality-provider/pull/179) Change `btc_pk` text to `eots_pk` in CLI
 * [#182](https://github.com/babylonlabs-io/finality-provider/pull/182) Remove fp manager
 * [#184](https://github.com/babylonlabs-io/finality-provider/pull/184) eots manager sign record store
+* [#189](https://github.com/babylonlabs-io/finality-provider/pull/189) Remove `fpd register-finality-provider` cmd
 * [#190](https://github.com/babylonlabs-io/finality-provider/pull/190) Benchmark pub rand
 
 ### Bug Fixes